Numeric.Floating.IEEE

Description

This module provides IEEE 754-compliant operations for floating-point numbers.

The functions in this module assume that the given floating-point type conform to IEEE 754 format.

Since RealFloat constraint is insufficient to query properties of a NaN, the functions here assumes all NaN as positive, quiet. If you want better treatment for NaNs, use the module Numeric.Floating.IEEE.NaN.

Since floating-point exceptions cannot be accessed from Haskell in normal way, the operations provided by this module ignore exceptional behavior. Don't let fp exceptions trap.

On i386 target, you may need to set -msse2 option to get correct floating-point behavior.

5.3 Homogeneous general-computational operations

5.3.1 General operations

round' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTiesToEven operation.

\(x :: Double) -> isFinite x ==> (round' x == fromInteger (round x))

>>> round' (-0.5)
-0.0

roundAway' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTiesToAway operation.

\(x :: Double) -> isFinite x ==> roundAway' x == fromInteger (roundAway x)

>>> roundAway' (-0.5)
-1.0
>>> roundAway' (-0.4)
-0.0

truncate' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTowardZero operation.

\(x :: Double) -> isFinite x ==> truncate' x == fromInteger (truncate x)

>>> truncate' (-0.5)
-0.0

ceiling' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTowardPositive operation.

\(x :: Double) -> isFinite x ==> ceiling' x == fromInteger (ceiling x)

>>> ceiling' (-0.8)
-0.0
>>> ceiling' (-0.5)
-0.0

floor' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTowardNegative operation.

\(x :: Double) -> isFinite x ==> floor' x == fromInteger (floor x)

>>> floor' (-0.1)
-1.0
>>> floor' (-0)
-0.0

nextUp :: RealFloat a => a -> a

IEEE 754 nextUp operation.

>>> nextUp 1 == (0x1.000002p0 :: Float)
True
>>> nextUp 1 == (0x1.0000_0000_0000_1p0 :: Double)
True
>>> nextUp (1/0) == (1/0 :: Double)
True
>>> nextUp (-1/0) == (- maxFinite :: Double)
True
>>> nextUp 0 == (0x1p-1074 :: Double)
True
>>> nextUp (-0) == (0x1p-1074 :: Double)
True
>>> nextUp (-0x1p-1074) :: Double -- should be negative zero
-0.0

nextDown :: RealFloat a => a -> a

IEEE 754 nextDown operation.

>>> nextDown 1 == (0x1.ffff_ffff_ffff_fp-1 :: Double)
True
>>> nextDown 1 == (0x1.fffffep-1 :: Float)
True
>>> nextDown (1/0) == (maxFinite :: Double)
True
>>> nextDown (-1/0) == (-1/0 :: Double)
True
>>> nextDown 0 == (-0x1p-1074 :: Double)
True
>>> nextDown (-0) == (-0x1p-1074 :: Double)
True
>>> nextDown 0x1p-1074 -- should be positive zero
0.0
>>> nextDown 0x1p-1022 == (0x0.ffff_ffff_ffff_fp-1022 ::Double)
True

nextTowardZero is not in IEEE, but may be useful to some.

nextTowardZero :: RealFloat a => a -> a

>>> nextTowardZero 1 == (0x1.ffff_ffff_ffff_fp-1 :: Double)
True
>>> nextTowardZero 1 == (0x1.fffffep-1 :: Float)
True
>>> nextTowardZero (1/0) == (maxFinite :: Double)
True
>>> nextTowardZero (-1/0) == (-maxFinite :: Double)
True
>>> nextTowardZero 0 :: Double -- should be positive zero
0.0
>>> nextTowardZero (-0 :: Double) -- should be negative zero
-0.0
>>> nextTowardZero 0x1p-1074 :: Double
0.0

remainder :: RealFloat a => a -> a -> a

IEEE 754 remainder operation.

\(r=x-yn\), where n is the integer nearest the exact number \(n=\mathrm{round}(x/y)\)

5.3.2 Decimal operations (not supported)

Not supported.

5.3.3 logBFormat operations

scaleFloatTiesToEven :: RealFloat a => Int -> a -> a

IEEE 754 scaleB operation, with each rounding attributes.

scaleFloatTiesToAway :: RealFloat a => Int -> a -> a

IEEE 754 scaleB operation, with each rounding attributes.

scaleFloatTowardPositive :: RealFloat a => Int -> a -> a

IEEE 754 scaleB operation, with each rounding attributes.

scaleFloatTowardNegative :: RealFloat a => Int -> a -> a

IEEE 754 scaleB operation, with each rounding attributes.

scaleFloatTowardZero :: RealFloat a => Int -> a -> a

IEEE 754 scaleB operation, with each rounding attributes.

The Haskell counterpart for logB operation is exponent. Note that logB and exponent are different by one: logB x = exponent x - 1

exponent :: RealFloat a => a -> Int

exponent corresponds to the second component of decodeFloat. exponent 0 = 0 and for finite nonzero x, exponent x = snd (decodeFloat x) + floatDigits x. If x is a finite floating-point number, it is equal in value to significand x * b ^^ exponent x, where b is the floating-point radix. The behaviour is unspecified on infinite or NaN values.

5.4 formatOf general-computational operations

5.4.1 Arithmetic operations

For IEEE-compliant floating-point types, (+), (-), (*), (/), and sqrt from Prelude should be correctly-rounding. fusedMultiplyAdd is provided by this library. This library also provides "generic" version of the arithmetic operations, which can be useful if the target type is narrower than source.

(+) :: Num a => a -> a -> a infixl 6

(-) :: Num a => a -> a -> a infixl 6

(*) :: Num a => a -> a -> a infixl 7

(/) :: Fractional a => a -> a -> a infixl 7

Fractional division.

sqrt :: Floating a => a -> a

fusedMultiplyAdd :: RealFloat a => a -> a -> a -> a

IEEE 754 fusedMultiplyAdd operation.

May make use of hardware FMA instructions if the target architecture has it; set fma3 package flag on x86 systems.

\(a :: Double) (b :: Double) (c :: Double) -> fusedMultiplyAdd a b c == fromRational (toRational a * toRational b + toRational c)

genericAdd :: (RealFloat a, RealFloat b) => a -> a -> b infixl 6

IEEE 754 addition operation.

genericSub :: (RealFloat a, RealFloat b) => a -> a -> b infixl 6

IEEE 754 subtraction operation.

genericMul :: (RealFloat a, RealFloat b) => a -> a -> b infixl 7

IEEE 754 multiplication operation.

genericDiv :: (RealFloat a, RealFloat b) => a -> a -> b infixl 7

IEEE 754 division operation.

genericSqrt: not implemented yet

genericFusedMultiplyAdd :: (RealFloat a, RealFloat b) => a -> a -> a -> b

IEEE 754 fusedMultiplyAdd operation.

The rounding direction of fromInteger for the standard floating-point types may vary depending on the magnitude of input. To get consistent rounding behavior, fromIntegerTiesToEven can be used instead.

fromIntegerTiesToEven :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegerTiesToAway :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegerTowardPositive :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegerTowardNegative :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegerTowardZero :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegralTiesToEven :: (Integral i, RealFloat a) => i -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegralTiesToAway :: (Integral i, RealFloat a) => i -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegralTowardPositive :: (Integral i, RealFloat a) => i -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegralTowardNegative :: (Integral i, RealFloat a) => i -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegralTowardZero :: (Integral i, RealFloat a) => i -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromRationalTiesToEven :: RealFloat a => Rational -> a

fromRationalTiesToAway :: RealFloat a => Rational -> a

fromRationalTowardPositive :: RealFloat a => Rational -> a

fromRationalTowardNegative :: RealFloat a => Rational -> a

fromRationalTowardZero :: RealFloat a => Rational -> a

For IEEE-compliant floating-point types, round, truncate, ceiling, and floor from RealFrac class should be correct implementation of IEEE 754 convertToInteger operations. To complete them, roundAway is provided by this library. Note that Haskell's round is specified to be ties-to-even, whereas C's round is ties-to-away.

round :: (RealFrac a, Integral b) => a -> b

round x returns the nearest integer to x; the even integer if x is equidistant between two integers

roundAway :: (RealFrac a, Integral b) => a -> b

IEEE 754 convertToIntegerTiesToAway operation.

>>> roundAway 4.5
5

truncate :: (RealFrac a, Integral b) => a -> b

truncate x returns the integer nearest x between zero and x

ceiling :: (RealFrac a, Integral b) => a -> b

ceiling x returns the least integer not less than x

floor :: (RealFrac a, Integral b) => a -> b

floor x returns the greatest integer not greater than x

5.4.2 Conversion operations for floating-point formats and decimal character sequences

Unfortunately, realToFrac does not have a good semantics, and changes behavior with rewrite rules (consider realToFrac (0/0 :: Float) :: Double). As an alternative, this library provides realFloatToFrac, with well-defined semantics on signed zeroes, infinities and NaNs. Like realToFrac, realFloatToFrac comes with some rewrite rules for particular types, but they should not change behavior.

realFloatToFrac :: (RealFloat a, Fractional b) => a -> b

Similar to realToFrac, but treats NaN, infinities, negative zero even if the rewrite rule is off.

canonicalize :: RealFloat a => a -> a

convertFromDecimalCharacter: not implemented.

convertToDecimalCharacter: not implemented.

5.4.3 Conversion operations for binary formats

convertFromHexCharacter: not implemented.

convertToHexCharacter: not implemented.

5.5 Quiet-computational operations

5.5.1 Sign bit operations

For IEEE-compliant floating-point types, negate and abs from Prelude should comply with IEEE semantics.

negate :: Num a => a -> a

Unary negation.

abs :: Num a => a -> a

Absolute value.

copySign: not implemented

5.5.2 Decimal re-encoding operations (not supported)

Not supported.

5.6 Signaling-computational operations

5.6.1 Comparisons (not supported)

This library does not support floating-point exceptions.

5.7 Non-computational operations

5.7.1 Conformance predicates (not supported)

Not supported.

5.7.2 General operations

Functions in this module disregards the content of NaNs: sign bit, signaling-or-quiet, and payload. All NaNs are treated as quiet, positive. To properly handle NaNs, use the typeclass and functions from Numeric.Floating.IEEE.NaN.

data Class

Constructors

SignalingNaN

QuietNaN

NegativeInfinity

NegativeNormal

NegativeSubnormal

NegativeZero

PositiveZero

PositiveSubnormal

PositiveNormal

PositiveInfinity

classify :: RealFloat a => a -> Class

Treats NaNs as quiet

isSignMinus :: RealFloat a => a -> Bool

IEEE 754 isSignMinus operation.

Since RealFloat constraint is insufficient to query the sign of NaNs, this function treats all NaNs as positive.

isNormal :: RealFloat a => a -> Bool

IEEE 754 isNormal operation.

isFinite :: RealFloat a => a -> Bool

IEEE 754 isFinite operation.

isZero :: RealFloat a => a -> Bool

IEEE 754 isZero operation.

isDenormalized :: RealFloat a => a -> Bool

True if the argument is too small to be represented in normalized format

isInfinite :: RealFloat a => a -> Bool

True if the argument is an IEEE infinity or negative infinity

isNaN :: RealFloat a => a -> Bool

True if the argument is an IEEE "not-a-number" (NaN) value

isSignaling: not supported here. isCanonical: not supported.

floatRadix :: RealFloat a => a -> Integer

a constant function, returning the radix of the representation (often 2)

compareByTotalOrder :: RealFloat a => a -> a -> Ordering

Comparison with IEEE 754 totalOrder predicate.

Since RealFloat constraint is insufficient to query the sign and payload of NaNs, this function treats all NaNs as positive.

Floating-point numbers are ordered as, \(-\infty < \text{negative reals} < -0 < +0 < \text{positive reals} < +\infty < \mathrm{NaN}\).

compareByTotalOrderMag :: RealFloat a => a -> a -> Ordering

Comparison with IEEE 754 totalOrderMag predicate.

5.7.3 Decimal operation (not supported)

Not supported.

5.7.4 Operations on subsets of flags (not supported)

Not supported.

9. Recommended operations

9.5 Augmented arithmetic operations

augmentedAddition :: RealFloat a => a -> a -> (a, a)

IEEE 754 augmentedAddition operation.

augmentedSubtraction :: RealFloat a => a -> a -> (a, a)

IEEE 754 augmentedSubtraction operation.

augmentedMultiplication :: RealFloat a => a -> a -> (a, a)

IEEE 754 augmentedMultiplication operation.

9.6 Minimum and maximum operations

minimum' :: RealFloat a => a -> a -> a

IEEE 754 minimum operation. -0 is smaller than +0. Propagates NaNs.

minimumNumber :: RealFloat a => a -> a -> a

IEEE 754 minimumNumber operation. -0 is smaller than +0. Treats NaNs as missing data.

maximum' :: RealFloat a => a -> a -> a

IEEE 754 maximum operation. -0 is smaller than +0. Propagates NaNs.

maximumNumber :: RealFloat a => a -> a -> a

IEEE 754 maximumNumber operation. -0 is smaller than +0. Treats NaNs as missing data.

minimumMagnitude :: RealFloat a => a -> a -> a

IEEE 754 minimumMagnitude operation.

minimumMagnitudeNumber :: RealFloat a => a -> a -> a

IEEE 754 minimumMagnitudeNumber operation.

maximumMagnitude :: RealFloat a => a -> a -> a

IEEE 754 maximumMagnitude operation.

maximumMagnitudeNumber :: RealFloat a => a -> a -> a

IEEE 754 maximumMagnitudeNumber operation.

Floating-point constants

minPositive :: RealFloat a => a

>>> (minPositive :: Float) == 0x1p-149
True
>>> (minPositive :: Double) == 0x1p-1074
True
>>> nextDown (minPositive :: Float)
0.0
>>> nextDown (minPositive :: Double)
0.0

minPositiveNormal :: RealFloat a => a

>>> (minPositiveNormal :: Float) == 0x1p-126
True
>>> (minPositiveNormal :: Double) == 0x1p-1022
True
>>> isDenormalized (minPositiveNormal :: Float)
False
>>> isDenormalized (minPositiveNormal :: Double)
False
>>> isDenormalized (nextDown (minPositiveNormal :: Float))
True
>>> isDenormalized (nextDown (minPositiveNormal :: Double))
True

maxFinite :: RealFloat a => a

>>> (maxFinite :: Float) == 0x1.fffffep+127
True
>>> (maxFinite :: Double) == 0x1.ffff_ffff_ffff_fp+1023
True