Numeric.Floating.IEEE.Internal

Documentation

truncate :: (RealFrac a, Integral b) => a -> b

truncate x returns the integer nearest x between zero and x

round :: (RealFrac a, Integral b) => a -> b

round x returns the nearest integer to x; the even integer if x is equidistant between two integers

ceiling :: (RealFrac a, Integral b) => a -> b

ceiling x returns the least integer not less than x

floor :: (RealFrac a, Integral b) => a -> b

floor x returns the greatest integer not greater than x

isFloatBinary32 :: Bool

isDoubleBinary64 :: Bool

minPositive :: RealFloat a => a

(minPositive :: Float) == 0x1p-149

(minPositive :: Double) == 0x1p-1074

nextDown (minPositive :: Float) == 0

nextDown (minPositive :: Double) == 0

minPositiveNormal :: RealFloat a => a

(minPositiveNormal :: Float) == 0x1p-126

(minPositiveNormal :: Double) == 0x1p-1022

not (isDenormalized (minPositiveNormal :: Float))

not (isDenormalized (minPositiveNormal :: Double))

isDenormalized (nextDown (minPositiveNormal :: Float))

isDenormalized (nextDown (minPositiveNormal :: Double))

maxFinite :: RealFloat a => a

(maxFinite :: Float) == 0x1.fffffep+127

(maxFinite :: Double) == 0x1.ffff_ffff_ffff_fp+1023

negateIntAsWord :: Int -> Word

negateIntAsWord minBound == fromInteger (negate (fromIntegral (minBound :: Int)))

absIntAsWord :: Int -> Word

absIntAsWord minBound == fromInteger (abs (fromIntegral (minBound :: Int)))

data Class

Constructors

SignalingNaN

QuietNaN

NegativeInfinity

NegativeNormal

NegativeSubnormal

NegativeZero

PositiveZero

PositiveSubnormal

PositiveNormal

PositiveInfinity

isNormal :: RealFloat a => a -> Bool

IEEE 754 isNormal operation.

isFloatNormal :: Float -> Bool

isDoubleNormal :: Double -> Bool

isFinite :: RealFloat a => a -> Bool

IEEE 754 isFinite operation.

isZero :: RealFloat a => a -> Bool

IEEE 754 isZero operation.

isSignMinus :: RealFloat a => a -> Bool

IEEE 754 isSignMinus operation.

Since RealFloat constraint is insufficient to query the sign of NaNs, this function treats all NaNs as positive.

compareByTotalOrder :: RealFloat a => a -> a -> Ordering

Comparison with IEEE 754 totalOrder predicate.

Since RealFloat constraint is insufficient to query the sign and payload of NaNs, this function treats all NaNs as positive.

Floating-point numbers are ordered as, \(-\infty < \text{negative reals} < -0 < +0 < \text{positive reals} < +\infty < \mathrm{NaN}\).

compareByTotalOrderMag :: RealFloat a => a -> a -> Ordering

Comparison with IEEE 754 totalOrderMag predicate.

classify :: RealFloat a => a -> Class

Treats NaNs as quiet

classifyFloat :: Float -> Class

classifyDouble :: Double -> Class

realFloatToFrac :: (RealFloat a, Fractional b) => a -> b

Similar to realToFrac, but treats NaN, infinities, negative zero even if the rewrite rule is off.

The properties of NaN may or may not be kept.

minimum' :: RealFloat a => a -> a -> a

IEEE 754 minimum operation. -0 is smaller than +0. Propagates NaNs.

minimumNumber :: RealFloat a => a -> a -> a

IEEE 754 minimumNumber operation. -0 is smaller than +0. Treats NaNs as missing data.

maximum' :: RealFloat a => a -> a -> a

IEEE 754 maximum operation. -0 is smaller than +0. Propagates NaNs.

maximumNumber :: RealFloat a => a -> a -> a

IEEE 754 maximumNumber operation. -0 is smaller than +0. Treats NaNs as missing data.

minimumMagnitude :: RealFloat a => a -> a -> a

IEEE 754 minimumMagnitude operation.

minimumMagnitudeNumber :: RealFloat a => a -> a -> a

IEEE 754 minimumMagnitudeNumber operation.

maximumMagnitude :: RealFloat a => a -> a -> a

IEEE 754 maximumMagnitude operation.

maximumMagnitudeNumber :: RealFloat a => a -> a -> a

IEEE 754 maximumMagnitudeNumber operation.

nextUp :: RealFloat a => a -> a

IEEE 754 nextUp operation.

nextUp 1 == (0x1.000002p0 :: Float)

nextUp 1 == (0x1.0000_0000_0000_1p0 :: Double)

nextUp (1/0) == (1/0 :: Double)

nextUp (-1/0) == (- maxFinite :: Double)

nextUp 0 == (0x1p-1074 :: Double)

nextUp (-0) == (0x1p-1074 :: Double)

nextUp (-0x1p-1074) == (-0 :: Double)

isNegativeZero (nextUp (-0x1p-1074) :: Double)

nextDown :: RealFloat a => a -> a

IEEE 754 nextDown operation.

nextDown 1 == (0x1.ffff_ffff_ffff_fp-1 :: Double)

nextDown 1 == (0x1.fffffep-1 :: Float)

nextDown (1/0) == (maxFinite :: Double)

nextDown (-1/0) == (-1/0 :: Double)

nextDown 0 == (-0x1p-1074 :: Double)

nextDown (-0) == (-0x1p-1074 :: Double)

nextDown 0x1p-1074 == (0 :: Double)

nextDown 0x1p-1022 == (0x0.ffff_ffff_ffff_fp-1022 ::Double)

nextTowardZero :: RealFloat a => a -> a

nextTowardZero 1 == (0x1.ffff_ffff_ffff_fp-1 :: Double)

nextTowardZero 1 == (0x1.fffffep-1 :: Float)

nextTowardZero (1/0) == (maxFinite :: Double)

nextTowardZero (-1/0) == (-maxFinite :: Double)

nextTowardZero 0 == (0 :: Double)

isNegativeZero (nextTowardZero (-0 :: Double))

nextTowardZero 0x1p-1074 == (0 :: Double)

nextUp_positive :: RealFloat a => a -> a

nextDown_positive :: RealFloat a => a -> a

nextUpFloat :: Float -> Float

nextUpFloat 1 == 0x1.000002p0

nextUpFloat (1/0) == 1/0

nextUpFloat (-1/0) == - maxFinite

nextUpFloat 0 == 0x1p-149

nextUpFloat (-0) == 0x1p-149

isNegativeZero (nextUpFloat (-0x1p-149))

nextUpDouble :: Double -> Double

nextUpDouble 1 == 0x1.0000_0000_0000_1p0

nextUpDouble (1/0) == 1/0

nextUpDouble (-1/0) == - maxFinite

nextUpDouble 0 == 0x1p-1074

nextUpDouble (-0) == 0x1p-1074

isNegativeZero (nextUpDouble (-0x1p-1074))

nextDownFloat :: Float -> Float

nextDownFloat 1 == 0x1.fffffep-1

nextDownFloat (1/0) == maxFinite

nextDownFloat (-1/0) == -1/0

nextDownFloat 0 == -0x1p-149

nextDownFloat (-0) == -0x1p-149

nextDownFloat 0x1p-149 == 0

nextDownDouble :: Double -> Double

nextDownDouble 1 == 0x1.ffff_ffff_ffff_fp-1

nextDownDouble (1/0) == maxFinite

nextDownDouble (-1/0) == -1/0

nextDownDouble 0 == -0x1p-1074

nextDownDouble (-0) == -0x1p-1074

nextDownDouble 0x1p-1074 == 0

nextTowardZeroFloat :: Float -> Float

nextTowardZeroFloat 1 == 0x1.fffffep-1

nextTowardZeroFloat (-1) == -0x1.fffffep-1

nextTowardZeroFloat (1/0) == maxFinite

nextTowardZeroFloat (-1/0) == -maxFinite

nextTowardZeroFloat 0 == 0

isNegativeZero (nextTowardZeroFloat (-0))

nextTowardZeroFloat 0x1p-149 == 0

nextTowardZeroDouble :: Double -> Double

nextTowardZeroDouble 1 == 0x1.ffff_ffff_ffff_fp-1

nextTowardZeroDouble (-1) == -0x1.ffff_ffff_ffff_fp-1

nextTowardZeroDouble (1/0) == maxFinite

nextTowardZeroDouble (-1/0) == -maxFinite

nextTowardZeroDouble 0 == 0

isNegativeZero (nextTowardZeroDouble (-0))

nextTowardZeroDouble 0x1p-1074 == 0

isMantissaEven :: RealFloat a => a -> Bool

twoSum :: Num a => a -> a -> (a, a)

\(a :: Double) (b :: Double) -> let (x, y) = twoSum a b in a + b == x && toRational a + toRational b == toRational x + toRational y

add_roundToOdd :: RealFloat a => a -> a -> a

split :: RealFloat a => a -> (a, a)

twoProduct_generic :: RealFloat a => a -> a -> (a, a)

twoProductFloat_viaDouble :: Float -> Float -> (Float, Float)

twoProduct :: RealFloat a => a -> a -> (a, a)

\(a :: Double) (b :: Double) -> let (x, y) = twoProduct a b in a * b == x && fromRational (toRational a * toRational b - toRational x) == y

twoProduct_nonscaling :: RealFloat a => a -> a -> (a, a)

fusedMultiplyAdd_twoProduct :: RealFloat a => a -> a -> a -> a

fusedMultiplyAddFloat_viaDouble :: Float -> Float -> Float -> Float

fusedMultiplyAdd_viaInteger :: RealFloat a => a -> a -> a -> a

fusedMultiplyAdd_viaRational :: RealFloat a => a -> a -> a -> a

fusedMultiplyAdd :: RealFloat a => a -> a -> a -> a

IEEE 754 fusedMultiplyAdd operation.

\(a :: Double) (b :: Double) (c :: Double) -> fusedMultiplyAdd a b c == fromRational (toRational a * toRational b + toRational c)

genericAdd :: (RealFloat a, RealFloat b) => a -> a -> b infixl 6

IEEE 754 addition operation.

genericSub :: (RealFloat a, RealFloat b) => a -> a -> b infixl 6

IEEE 754 subtraction operation.

genericMul :: (RealFloat a, RealFloat b) => a -> a -> b infixl 7

IEEE 754 multiplication operation.

genericDiv :: (RealFloat a, RealFloat b) => a -> a -> b infixl 7

IEEE 754 division operation.

genericFusedMultiplyAdd :: (RealFloat a, RealFloat b) => a -> a -> a -> b

IEEE 754 fusedMultiplyAdd operation.

isSubFloatingType :: (RealFloat a, RealFloat b) => a -> b -> Bool

Returns True if a is a subtype of b

>>> isSubFloatingType (undefined :: Float) (undefined :: Double)
True
>>> isSubFloatingType (undefined :: Double) (undefined :: Float)
False
>>> isSubFloatingType (undefined :: Double) (undefined :: Double)
True

isSubFloatingTypeProxy :: (RealFloat a, RealFloat b) => Proxy a -> Proxy b -> Bool

Returns True if a is a subtype of b

>>> isSubFloatingTypeProxy (Proxy :: Proxy Float) (Proxy :: Proxy Double)
True
>>> isSubFloatingTypeProxy (Proxy :: Proxy Double) (Proxy :: Proxy Float)
False
>>> isSubFloatingTypeProxy (Proxy :: Proxy Double) (Proxy :: Proxy Double)
True

remainder :: RealFloat a => a -> a -> a

IEEE 754 remainder operation.

round' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTiesToEven operation.

\(x :: Double) -> isFinite x ==> (round' x == fromInteger (round x))

>>> round' (-0.5)
-0.0

roundAway' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTiesToAway operation.

\(x :: Double) -> isFinite x ==> roundAway' x == fromInteger (roundAway x)

>>> roundAway' (-0.5)
-1.0
>>> roundAway' (-0.4)
-0.0

truncate' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTowardZero operation.

\(x :: Double) -> isFinite x ==> truncate' x == fromInteger (truncate x)

>>> truncate' (-0.5)
-0.0

ceiling' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTowardPositive operation.

\(x :: Double) -> isFinite x ==> ceiling' x == fromInteger (ceiling x)

>>> ceiling' (-0.8)
-0.0
>>> ceiling' (-0.5)
-0.0

floor' :: RealFloat a => a -> a

IEEE 754 roundToIntegralTowardNegative operation.

\(x :: Double) -> isFinite x ==> floor' x == fromInteger (floor x)

>>> floor' (-0.1)
-1.0
>>> floor' (-0)
-0.0

roundAway :: (RealFrac a, Integral b) => a -> b

IEEE 754 convertToIntegerTiesToAway operation.

>>> roundAway 4.5
5

newtype Exactness a

Constructors

Exactness

Fields

isExact :: Bool

newtype RoundToOdd a

Constructors

RoundToOdd

Fields

roundToOdd :: a

newtype RoundTiesTowardZero a

Constructors

RoundTiesTowardZero

Fields

roundTiesTowardZero :: a

newtype RoundTowardZero a

Constructors

RoundTowardZero

Fields

roundTowardZero :: a

newtype RoundTowardNegative a

Constructors

RoundTowardNegative

Fields

roundTowardNegative :: a

newtype RoundTowardPositive a

Constructors

RoundTowardPositive

Fields

roundTowardPositive :: a

newtype RoundTiesToAway a

Constructors

RoundTiesToAway

Fields

roundTiesToAway :: a

newtype RoundTiesToEven a

Constructors

RoundTiesToEven

Fields

roundTiesToEven :: a

class Functor f => RoundingStrategy f where

Methods

exact :: a -> f a

inexact

Arguments

:: Ordering
LT -> toward-zero is the nearest, EQ -> midpoint, GT -> away-from-zero is the nearest

-> Bool
negative (True -> negative, False -> positive)

-> Int
parity (even -> toward-zero is even, odd -> toward-zero is odd)

-> a
toward zero

-> a
away from zero

-> f a

quotRemByExpt :: Integer -> Integer -> Int -> (Integer, Integer)

multiplyByExpt :: Integer -> Integer -> Int -> Integer

divideByExpt :: Integer -> Integer -> Int -> Integer

fromIntegerR :: (RealFloat a, RoundingStrategy f) => Integer -> f a

fromIntegerTiesToEven :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegerTiesToAway :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegerTowardPositive :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegerTowardNegative :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegerTowardZero :: RealFloat a => Integer -> a

IEEE 754 convertFromInt operation, with each rounding attributes.

fromIntegralR :: (Integral i, RealFloat a, RoundingStrategy f) => i -> f a

fromIntegralTiesToEven :: (Integral i, RealFloat a) => i -> a

fromIntegralTiesToAway :: (Integral i, RealFloat a) => i -> a

fromIntegralTowardPositive :: (Integral i, RealFloat a) => i -> a

fromIntegralTowardNegative :: (Integral i, RealFloat a) => i -> a

fromIntegralTowardZero :: (Integral i, RealFloat a) => i -> a

fromIntegralRBits :: (Integral i, Bits i, RealFloat a, RoundingStrategy f) => i -> f a

boundsForExactConversion :: (Integral i, Bits i, RealFloat a) => Proxy a -> (Maybe i, Maybe i)

>>> boundsForExactConversion (Proxy :: Proxy Double) :: (Maybe Integer, Maybe Integer) -- (Just (-2^53),Just (2^53))
(Just (-9007199254740992),Just 9007199254740992)
>>> boundsForExactConversion (Proxy :: Proxy Double) :: (Maybe Int32, Maybe Int32) -- the conversion is always exact
(Nothing,Nothing)
>>> boundForExactConversion (Proxy :: Proxy Float) :: (Maybe Word, Maybe Word) -- (Nothing,Just (2^23))
(Nothing,Just 8388608)

pureIfThenElse :: Bool -> a -> a -> a

positiveWordToBinaryFloatR :: (RealFloat a, RoundingStrategy f) => Bool -> Word -> f a

positiveWordToBinaryFloatR# :: (RealFloat a, RoundingStrategy f) => Bool -> Word# -> f a

fromPositiveIntegerR :: (RealFloat a, RoundingStrategy f) => Bool -> Integer -> f a

fromRationalR :: (RealFloat a, RoundingStrategy f) => Rational -> f a

fromRationalTiesToEven :: RealFloat a => Rational -> a

fromRationalTiesToAway :: RealFloat a => Rational -> a

fromRationalTowardPositive :: RealFloat a => Rational -> a

fromRationalTowardNegative :: RealFloat a => Rational -> a

fromRationalTowardZero :: RealFloat a => Rational -> a

fromRatioR

Arguments

:: (RealFloat a, RoundingStrategy f)

=> Integer
numerator

-> Integer
denominator

-> f a

fromPositiveRatioR

Arguments

:: (RealFloat a, RoundingStrategy f)

=> Bool
True if the result will be negated

-> Integer
numerator (> 0)

-> Integer
denominator (> 0)

-> f a

encodeFloatR :: (RealFloat a, RoundingStrategy f) => Integer -> Int -> f a

encodeFloatTiesToEven :: RealFloat a => Integer -> Int -> a

encodeFloatTiesToAway :: RealFloat a => Integer -> Int -> a

encodeFloatTowardPositive :: RealFloat a => Integer -> Int -> a

encodeFloatTowardNegative :: RealFloat a => Integer -> Int -> a

encodeFloatTowardZero :: RealFloat a => Integer -> Int -> a

encodePositiveFloatR :: (RealFloat a, RoundingStrategy f) => Bool -> Integer -> Int -> f a

encodePositiveFloatR# :: (RealFloat a, RoundingStrategy f) => Bool -> Integer -> Int# -> f a

augmentedAddition :: RealFloat a => a -> a -> (a, a)

IEEE 754 augmentedAddition operation.

augmentedSubtraction :: RealFloat a => a -> a -> (a, a)

IEEE 754 augmentedSubtraction operation.

augmentedMultiplication :: RealFloat a => a -> a -> (a, a)

IEEE 754 augmentedMultiplication operation.

augmentedAddition_viaRational :: (RealFloat a, Show a) => a -> a -> (a, a)

augmentedMultiplication_viaRational :: (RealFloat a, Show a) => a -> a -> (a, a)