Numeric.Floating.IEEE.NaN

Documentation

class RealFloat a => RealFloatNaN a where

An instance of this class supports manipulation of NaN.

Minimal complete definition

(copySign | isSignMinus), (isSignaling | classify), getPayload, setPayload, setPayloadSignaling

Methods

copySign :: a -> a -> a

Returns the first operand, with the sign of the second.

IEEE 754 copySign operation.

isSignMinus :: a -> Bool

Returns True if the operand is a negative number, negative infinity, negative zero, or a NaN with negative sign bit.

IEEE 754 isSignMinus operation.

isSignaling :: a -> Bool

Returns True if the operand is a signaling NaN.

IEEE 754 isSignaling operation.

getPayload :: a -> a

Returns the payload of a NaN. Returns -1 if the operand is not a NaN.

IEEE 754 getPayload operation.

setPayload :: a -> a

Returns a quiet NaN with a given payload. Returns a positive zero if the payload is invalid.

IEEE 754 setPayload operation.

setPayloadSignaling :: a -> a

Returns a signaling NaN with a given payload. Returns a positive zero if the payload is invalid.

IEEE 754 setPayloadSignaling operation.

classify :: a -> Class

IEEE 754 class operation.

equalByTotalOrder :: a -> a -> Bool

Equality with IEEE 754 totalOrder operation.

compareByTotalOrder :: a -> a -> Ordering

Comparison with IEEE 754 totalOrder operation.

data Class

Constructors

SignalingNaN

QuietNaN

NegativeInfinity

NegativeNormal

NegativeSubnormal

NegativeZero

PositiveZero

PositiveSubnormal

PositiveNormal

PositiveInfinity

newtype TotallyOrdered a

Constructors

TotallyOrdered a